Move cron validatioin methods into a mixin
This will allow us to use them in the other cron resource eventually, but definitely in the chef_client_cron resource today.
