+++ spec/unit/resource/cron_d_spec.rb
@@ -33,58 +33,4 @@ describe Chef::Resource::CronD do
   it "the cron_name property is the name_property" do
     expect(resource.cron_name).to eql("cronify")
   end
-
-  context "#validate_dow" do
-    it "it accepts a string day" do
-      expect(Chef::Resource::CronD.validate_dow("mon")).to be true
-    end
-
-    it "it accepts an integer day" do
-      expect(Chef::Resource::CronD.validate_dow(0)).to be true
-    end
-
-    it "it accepts the string of *" do
-      expect(Chef::Resource::CronD.validate_dow("*")).to be true
-    end
-
-    it "returns false for an out of range integer" do
-      expect(Chef::Resource::CronD.validate_dow(8)).to be false
-    end
-
-    it "returns false for an invalid string" do
-      expect(Chef::Resource::CronD.validate_dow("monday")).to be false
-    end
-  end
-
-  context "#validate_month" do
-    it "it accepts a string month" do
-      expect(Chef::Resource::CronD.validate_month("feb")).to be true
-    end
-
-    it "it accepts an integer month" do
-      expect(Chef::Resource::CronD.validate_month(2)).to be true
-    end
-
-    it "it accepts the string of *" do
-      expect(Chef::Resource::CronD.validate_month("*")).to be true
-    end
-
-    it "returns false for an out of range integer" do
-      expect(Chef::Resource::CronD.validate_month(13)).to be false
-    end
-
-    it "returns false for an invalid string" do
-      expect(Chef::Resource::CronD.validate_month("janurary")).to be false
-    end
-  end
-
-  context "#validate_numeric" do
-    it "returns true if the value is in the allowed range" do
-      expect(Chef::Resource::CronD.validate_numeric(5, 1, 100)).to be true
-    end
-
-    it "returns false if the value is out of the allowed range" do
-      expect(Chef::Resource::CronD.validate_numeric(-1, 1, 100)).to be false
-    end
-  end
 end
